Fix null ref in L0

+++ Tests-Legacy/L0/RunDistributedTests/_suite.ts
@@ -22,7 +22,9 @@ describe('Deploy Test Agent Suite', function () {
     });
 
     after(function () {
-        psr.kill();
+        if (psr) {
+            psr.kill();
+        }
     });
 
     if (psm.testSupported()) {
